feat: add specific error codes in EVM module
Add two error codes for when commands are blocked by a key rotation or signing in progress
