+++ utils/errors.go
@@ -1,0 +1,23 @@
+package utils
+
+import "github.com/cosmos/cosmos-sdk/types/errors"
+
+type causer interface {
+	Cause() error
+}
+
+// IsABCIError checks if the error is a (wrapped) registered error
+func IsABCIError(err error) bool {
+	for {
+		switch e := err.(type) {
+		case nil:
+			return false
+		case *errors.Error:
+			return true
+		case causer:
+			err = e.Cause()
+		default:
+			return false
+		}
+	}
+}
